hardware folder structure refactor - for clarity
