+++ controller/hardware/actuator/led/board-led.cpp
@@ -1,4 +1,4 @@
-#include <actuator/board-led.hpp>
+#include <led/board-led.hpp>
 #include <stm-hal/hal-gpio.hpp>
 
 // ------------------------------------------------------------------------------------
