+++ controller/hardware/actuator/led/board-neoled.cpp
@@ -1,4 +1,4 @@
-#include <actuator/board-neoled.hpp>
+#include <led/board-neoled.hpp>
 #include <system/system-neoled.hpp>
 
 // ------------------------------------------------------------------------------------
