+++ controller/hardware/actuator/motor/ev3-large-motor.cpp
@@ -1,4 +1,4 @@
-#include <actuator/nxt-motor.hpp>
+#include <motor/ev3-large-motor.hpp>
 #include <cstdint>
 
 // ------------------------------------------------------------------------------------
@@ -6,13 +6,13 @@
 // ------------------------------------------------------------------------------------
 
 /**
- * Create an NXT Motor.
+ * Create an EV3 Large Motor.
  *
  * @param *port pointer to motors port.
  * @param motor_pmw Timer Type value for motor pwm.
  * @param motor_encoder Timer Type value for motor encoder.
 */
-NXTMotor::NXTMotor(OutputPort *port, TimerType motor_pwm, TimerType motor_encoder)
+EV3LargeMotor::EV3LargeMotor(OutputPort *port, TimerType motor_pwm, TimerType motor_encoder)
 {
     m_port = port;
     m_motor_pwm = motor_pwm;
@@ -21,6 +21,8 @@ NXTMotor::NXTMotor(OutputPort *port, TimerType motor_pwm, TimerType motor_encode
     m_sync_motor = nullptr;
 
     m_synced = false;
+
+    m_motor_speed = DEFAULT_MOTOR_SPEED;
 }
 
 /**
@@ -28,11 +30,21 @@ NXTMotor::NXTMotor(OutputPort *port, TimerType motor_pwm, TimerType motor_encode
  *
  * @param rotation amount forward for the motor.
 */
-void NXTMotor::forward(int32_t rotation)
+void EV3LargeMotor::forward(int32_t rotation)
 {
-    drive_motor(0.0f, rotation, false);
+    if (rotation == 0)
+    {
+        return;
+    }
 
-    return;
+    if (rotation < 0)
+    {
+        backward(-rotation);
+        return;
+    }
+
+    float speed = m_motor_speed / 1000;
+    drive_motor(speed, rotation, false);
 }
 
 /**
@@ -40,27 +52,42 @@ void NXTMotor::forward(int32_t rotation)
  *
  * @param rotation amount backward for the motor.
 */
-void NXTMotor::backward(int32_t rotation)
+void EV3LargeMotor::backward(int32_t rotation)
 {
-    drive_motor(0.0f, -rotation, false);
+    if (rotation == 0)
+    {
+        return;
+    }
 
-    return;
+    if (rotation < 0)
+    {
+        forward(-rotation);
+        return;
+    }
+
+    float speed = m_motor_speed / 1000;
+    drive_motor(-speed, -rotation, false);
 }
 
 /**
  * Stop the motor.
 */
-void NXTMotor::stop()
+void EV3LargeMotor::stop()
 {
+    if (m_synced)
+    {
+        m_sync_motor->stop();
+    }
+
     return;
 }
 
 /**
- * Sync with another Motor.
+ * Sync with another Large Motor.
  *
  * @param motor address of other motor that is being synced with.
 */
-void NXTMotor::start_sync(NXTMotor *motor)
+void EV3LargeMotor::start_sync(EV3LargeMotor *motor)
 {
     if (m_synced)
     {
@@ -69,7 +96,6 @@ void NXTMotor::start_sync(NXTMotor *motor)
     }
 
     stop();
-    motor->stop();
 
     m_sync_motor = motor;
     m_synced = true;
@@ -80,7 +106,7 @@ void NXTMotor::start_sync(NXTMotor *motor)
 /**
  * Stops syncing the motors if is synced.
 */
-void NXTMotor::end_sync()
+void EV3LargeMotor::end_sync()
 {
     if (m_sync_motor == nullptr)
     {
@@ -98,7 +124,7 @@ void NXTMotor::end_sync()
  *
  * @return boolean - stalled value.
 */
-bool NXTMotor::is_stalled()
+bool EV3LargeMotor::is_stalled()
 {
     return false;
 }
@@ -109,7 +135,7 @@ bool NXTMotor::is_stalled()
  *
  * @return int32_t - tacho count.
 */
-int32_t NXTMotor::get_tacho_count()
+int32_t EV3LargeMotor::get_tacho_count()
 {
     return 0;
 }
@@ -119,7 +145,7 @@ int32_t NXTMotor::get_tacho_count()
  *
  * @param motor_pwm
 */
-void NXTMotor::set_motor_pwm(uint8_t motor_pwm)
+void EV3LargeMotor::set_motor_pwm(uint8_t motor_pwm)
 {
     m_motor_pwm = static_cast<TimerType>(motor_pwm);
 }
@@ -129,7 +155,7 @@ void NXTMotor::set_motor_pwm(uint8_t motor_pwm)
  *
  * @return motor_pwm
 */
-uint8_t NXTMotor::get_motor_pwm()
+uint8_t EV3LargeMotor::get_motor_pwm()
 {
     return static_cast<uint8_t>(m_motor_pwm);
 }
@@ -139,7 +165,7 @@ uint8_t NXTMotor::get_motor_pwm()
  *
  * @param motor_encoder
 */
-void NXTMotor::set_motor_encoder(uint8_t motor_encoder)
+void EV3LargeMotor::set_motor_encoder(uint8_t motor_encoder)
 {
     m_motor_encoder = static_cast<TimerType>(motor_encoder);
 }
@@ -150,7 +176,7 @@ void NXTMotor::set_motor_encoder(uint8_t motor_encoder)
  *
  * @return motor_encoder
 */
-uint8_t NXTMotor::get_motor_encoder()
+uint8_t EV3LargeMotor::get_motor_encoder()
 {
     return static_cast<uint8_t>(m_motor_encoder);
 }
@@ -162,9 +188,10 @@ uint8_t NXTMotor::get_motor_encoder()
  *
  * @param motor_speed
 */
-void NXTMotor::set_motor_speed(uint32_t motor_speed)
+void EV3LargeMotor::set_motor_speed(uint32_t motor_speed)
 {
     m_motor_speed = (motor_speed > MAX_MOTOR_SPEED) ? MAX_MOTOR_SPEED : motor_speed;
+    m_motor_speed = (motor_speed < MIN_MOTOR_SPEED) ? MIN_MOTOR_SPEED : motor_speed;
 }
 
 /**
@@ -172,7 +199,7 @@ void NXTMotor::set_motor_speed(uint32_t motor_speed)
  *
  * @return motor_speed
 */
-uint32_t NXTMotor::get_motor_speed(void)
+uint32_t EV3LargeMotor::get_motor_speed(void)
 {
     return m_motor_speed;
 }
@@ -188,11 +215,38 @@ uint32_t NXTMotor::get_motor_speed(void)
  * @param rotation amount of rotation of the motor.
  * @param immediate_return should we immediately return to program after telling the motor to run?
 */
-void NXTMotor::drive_motor(float speed, int32_t rotation, bool immediate_return)
+void EV3LargeMotor::drive_motor(float speed, int32_t rotation, bool immediate_return)
 {
     if (immediate_return);
 
-    if (speed);
+    if (!m_synced)
+    {
+        m_port->set_pwm_duty(speed);
 
-    if (rotation);
+        uint16_t now = m_port->get_encoder_ticks();
+        uint16_t encoder_ticks = now;
+
+        while (encoder_ticks + rotation < now)
+        {
+            m_port->update();
+
+            now = m_port->get_encoder_ticks();
+        }
+    }
+    else
+    {
+        m_port->set_pwm_duty(speed);
+        m_sync_motor->m_port->set_pwm_duty(speed);
+
+        uint16_t now = m_port->get_encoder_ticks();
+        uint16_t encoder_ticks = now;
+
+        while(encoder_ticks + rotation < now)
+        {
+            m_port->update();
+            m_sync_motor->m_port->update();
+
+            now = m_port->get_encoder_ticks();
+        }
+    }
 }
