+++ controller/hardware/actuator/motor/ev3-medium-motor.hpp
