+++ controller/hardware/actuator/motor/nxt-motor.cpp
@@ -1,4 +1,4 @@
-#include <actuator/ev3-large-motor.hpp>
+#include <motor/nxt-motor.hpp>
 #include <cstdint>
 
 // ------------------------------------------------------------------------------------
@@ -6,13 +6,13 @@
 // ------------------------------------------------------------------------------------
 
 /**
- * Create an EV3 Large Motor.
+ * Create an NXT Motor.
  *
  * @param *port pointer to motors port.
  * @param motor_pmw Timer Type value for motor pwm.
  * @param motor_encoder Timer Type value for motor encoder.
 */
-EV3LargeMotor::EV3LargeMotor(OutputPort *port, TimerType motor_pwm, TimerType motor_encoder)
+NXTMotor::NXTMotor(OutputPort *port, TimerType motor_pwm, TimerType motor_encoder)
 {
     m_port = port;
     m_motor_pwm = motor_pwm;
@@ -21,8 +21,6 @@ EV3LargeMotor::EV3LargeMotor(OutputPort *port, TimerType motor_pwm, TimerType mo
     m_sync_motor = nullptr;
 
     m_synced = false;
-
-    m_motor_speed = DEFAULT_MOTOR_SPEED;
 }
 
 /**
@@ -30,21 +28,11 @@ EV3LargeMotor::EV3LargeMotor(OutputPort *port, TimerType motor_pwm, TimerType mo
  *
  * @param rotation amount forward for the motor.
 */
-void EV3LargeMotor::forward(int32_t rotation)
+void NXTMotor::forward(int32_t rotation)
 {
-    if (rotation == 0)
-    {
-        return;
-    }
+    drive_motor(0.0f, rotation, false);
 
-    if (rotation < 0)
-    {
-        backward(-rotation);
-        return;
-    }
-
-    float speed = m_motor_speed / 1000;
-    drive_motor(speed, rotation, false);
+    return;
 }
 
 /**
@@ -52,42 +40,27 @@ void EV3LargeMotor::forward(int32_t rotation)
  *
  * @param rotation amount backward for the motor.
 */
-void EV3LargeMotor::backward(int32_t rotation)
+void NXTMotor::backward(int32_t rotation)
 {
-    if (rotation == 0)
-    {
-        return;
-    }
-
-    if (rotation < 0)
-    {
-        forward(-rotation);
-        return;
-    }
-
-    float speed = m_motor_speed / 1000;
-    drive_motor(-speed, -rotation, false);
-}
-
-/**
- * Stop the motor.
-*/
-void EV3LargeMotor::stop()
-{
-    if (m_synced)
-    {
-        m_sync_motor->stop();
-    }
+    drive_motor(0.0f, -rotation, false);
 
     return;
 }
 
 /**
- * Sync with another Large Motor.
+ * Stop the motor.
+*/
+void NXTMotor::stop()
+{
+    return;
+}
+
+/**
+ * Sync with another Motor.
  *
  * @param motor address of other motor that is being synced with.
 */
-void EV3LargeMotor::start_sync(EV3LargeMotor *motor)
+void NXTMotor::start_sync(NXTMotor *motor)
 {
     if (m_synced)
     {
@@ -96,6 +69,7 @@ void EV3LargeMotor::start_sync(EV3LargeMotor *motor)
     }
 
     stop();
+    motor->stop();
 
     m_sync_motor = motor;
     m_synced = true;
@@ -106,7 +80,7 @@ void EV3LargeMotor::start_sync(EV3LargeMotor *motor)
 /**
  * Stops syncing the motors if is synced.
 */
-void EV3LargeMotor::end_sync()
+void NXTMotor::end_sync()
 {
     if (m_sync_motor == nullptr)
     {
@@ -124,7 +98,7 @@ void EV3LargeMotor::end_sync()
  *
  * @return boolean - stalled value.
 */
-bool EV3LargeMotor::is_stalled()
+bool NXTMotor::is_stalled()
 {
     return false;
 }
@@ -135,7 +109,7 @@ bool EV3LargeMotor::is_stalled()
  *
  * @return int32_t - tacho count.
 */
-int32_t EV3LargeMotor::get_tacho_count()
+int32_t NXTMotor::get_tacho_count()
 {
     return 0;
 }
@@ -145,7 +119,7 @@ int32_t EV3LargeMotor::get_tacho_count()
  *
  * @param motor_pwm
 */
-void EV3LargeMotor::set_motor_pwm(uint8_t motor_pwm)
+void NXTMotor::set_motor_pwm(uint8_t motor_pwm)
 {
     m_motor_pwm = static_cast<TimerType>(motor_pwm);
 }
@@ -155,7 +129,7 @@ void EV3LargeMotor::set_motor_pwm(uint8_t motor_pwm)
  *
  * @return motor_pwm
 */
-uint8_t EV3LargeMotor::get_motor_pwm()
+uint8_t NXTMotor::get_motor_pwm()
 {
     return static_cast<uint8_t>(m_motor_pwm);
 }
@@ -165,7 +139,7 @@ uint8_t EV3LargeMotor::get_motor_pwm()
  *
  * @param motor_encoder
 */
-void EV3LargeMotor::set_motor_encoder(uint8_t motor_encoder)
+void NXTMotor::set_motor_encoder(uint8_t motor_encoder)
 {
     m_motor_encoder = static_cast<TimerType>(motor_encoder);
 }
@@ -176,7 +150,7 @@ void EV3LargeMotor::set_motor_encoder(uint8_t motor_encoder)
  *
  * @return motor_encoder
 */
-uint8_t EV3LargeMotor::get_motor_encoder()
+uint8_t NXTMotor::get_motor_encoder()
 {
     return static_cast<uint8_t>(m_motor_encoder);
 }
@@ -188,10 +162,9 @@ uint8_t EV3LargeMotor::get_motor_encoder()
  *
  * @param motor_speed
 */
-void EV3LargeMotor::set_motor_speed(uint32_t motor_speed)
+void NXTMotor::set_motor_speed(uint32_t motor_speed)
 {
     m_motor_speed = (motor_speed > MAX_MOTOR_SPEED) ? MAX_MOTOR_SPEED : motor_speed;
-    m_motor_speed = (motor_speed < MIN_MOTOR_SPEED) ? MIN_MOTOR_SPEED : motor_speed;
 }
 
 /**
@@ -199,7 +172,7 @@ void EV3LargeMotor::set_motor_speed(uint32_t motor_speed)
  *
  * @return motor_speed
 */
-uint32_t EV3LargeMotor::get_motor_speed(void)
+uint32_t NXTMotor::get_motor_speed(void)
 {
     return m_motor_speed;
 }
@@ -215,38 +188,11 @@ uint32_t EV3LargeMotor::get_motor_speed(void)
  * @param rotation amount of rotation of the motor.
  * @param immediate_return should we immediately return to program after telling the motor to run?
 */
-void EV3LargeMotor::drive_motor(float speed, int32_t rotation, bool immediate_return)
+void NXTMotor::drive_motor(float speed, int32_t rotation, bool immediate_return)
 {
     if (immediate_return);
 
-    if (!m_synced)
-    {
-        m_port->set_pwm_duty(speed);
+    if (speed);
 
-        uint16_t now = m_port->get_encoder_ticks();
-        uint16_t encoder_ticks = now;
-
-        while (encoder_ticks + rotation < now)
-        {
-            m_port->update();
-
-            now = m_port->get_encoder_ticks();
-        }
-    }
-    else
-    {
-        m_port->set_pwm_duty(speed);
-        m_sync_motor->m_port->set_pwm_duty(speed);
-
-        uint16_t now = m_port->get_encoder_ticks();
-        uint16_t encoder_ticks = now;
-
-        while(encoder_ticks + rotation < now)
-        {
-            m_port->update();
-            m_sync_motor->m_port->update();
-
-            now = m_port->get_encoder_ticks();
-        }
-    }
+    if (rotation);
 }
