+++ controller/hardware/actuator/motor/nxt-motor.hpp
