+++ controller/hardware/sensor/colour/ev3-colour-sensor.hpp
